+++ test/compute-engine/latex-syntax/serialize.test.ts
@@ -137,8 +137,8 @@ describe('LATEX', () => {
           [
             "ErrorCode",
             "'incompatible-domain'",
-            ["Domain", "Strings"],
-            ["Domain", "PositiveIntegers"]
+            "Strings",
+            "PositiveIntegers"
           ],
           22
         ]